Strategy Checkpoint: Combine Structs
Commit here, as a checkpoint with the current work.

The thought process is to combine the different `Clan` structs and
combine the `Player` structs. This would allow for a strategy of merging
two structs from different API calls and having all the information in
one struct that is `Clan`/`Player`.