Create HTTP Module
Created the HTTP module to isolate the behaviours of HTTP requests into
its own module, instead of having it all in the base module. Updated all
calls to reference the new module, without aliasing.

Created part of the `ExClash.Player` struct. I have identified the keys
that are basic types, and have commented out the keys that will require
additional parsing and formatting.