Finish Player, Add Hero, Equipment, Spells
Finished up the player struct, and parsing the response and getting the
content into structs. That includes the creation of the hero, hero
equipment, and spell structs.

Also, simplified the village and clan member role parsing so that each
module does not have to worry about updating the value. Now there is
pattern matching to check for those keys in a map, and will assign them
appropriately.