Conditionally Load Secrets
Only load the secret files if they exist. This moves their existence as
a requirement to run the application locally, or running tests. Just
won't be able to actually make API requests to Supercell without the
required token.

Also, tests still don't exist, kinda. There is only one to test the one
existing fucntion that exists in the base module.