Add First Unit Test
This includes adding the ability to pull a plug out of the passed in
query params. This allows for tests to not hit the network and actually
have a deterministic response independent of Supercell's response.

Also removed the unnecessary `auth/0` function, and just put the
authorization into the `Req.get/2` call.

Not sure how to word this, but I am interested in alternative stubbing
options. I would rather not have production code pattern match on an
existing key to know that it is not supposed to hit the network.