Player Details Test, Async Tests
Adding the player details test, since that is the only player based API
call. It was a quick win. There is lots to validate, however, the mere
matching of the tag is enough for me for now. This doesn't do much to
ensure proper struct formatting though.

Also made the tests asyncronous.