Add Search Test, Move `nil` Handling
Added the test for searching for a clan. Much like the other tests, this
is severely underasserted on. What it does test, is that the JSON parsed
response is put into the proper struct. But the keys and values are not
verified.

Also, a new strategy for handling lists and `nil` is to have the
struct's `format/1` handle it. Thus, each `format/1` should have a
pattern match for `nil`, and when `data` is a list.