Clan Capital Raid Seasons
This is a work in progress. Unfortunately, I just made the decision to
not layer my modules so deeply. Not that it is incorrect, but it is
obnoxious.

So, step 1; move `ExClash.Clan.Capital` to `ExClash.Capital`. Just
because capitals are only for clans, does not mean that I need the extra
layer in the modules. It is pretty clear, when talking about the capital
that it belongs to the clan.

Thanks for coming to my TED talk.